Fixed signed unsigned

+++ src/value_store.h
@@ -56,7 +56,7 @@ namespace FlatCircuit {
 
       unsigned long nextInd = simValueTable.size();
 
-      for (unsigned long i = 0; i < (int) bv.bitLength(); i++) {
+      for (unsigned long i = 0; i < (unsigned long) bv.bitLength(); i++) {
         simValueTable.push_back(bv.get(i));
       }
